Make `SpendAuth` signatures independent of the decaf377 hash-to-group mechanism.
Because the hash-to-group mechanism for decaf377 is still subject to change (we
need to select the quadratic nonresidue together with an optimized square root
algorithm), all uses of it are also subject to change.  Making `SpendAuth`
signatures independent of this evolution means that the (longer-term) spend
authorization keys are less likely to be invalidated.